pyproject.toml
[project]
name = "trading-bot"
version = "0.2.0"
description = "AI-Enhanced Crypto Trading Bot with MCP Integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core trading dependencies
    "python-binance>=1.0.19",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-telegram-bot>=20.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    # MCP and AI dependencies
    "aiohttp>=3.8.0",
    "asyncio-mqtt>=0.16.0",
    "websockets>=11.0.0",
    "aiofiles>=23.0.0",
    # Data analysis and visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    # News and sentiment analysis
    "feedparser>=6.0.10",
    "textblob>=0.17.1",
    "newspaper3k>=0.2.8",
    # Optional ML dependencies
    "scikit-learn>=1.3.0",
    "mcp-python>=0.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-asyncio>=0.21.0",
]

ml = [
    "xgboost>=1.7.0",
    "optuna>=3.3.0",
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
]

web = [
    "streamlit>=1.25.0",
    "dash>=2.12.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

all = [
    "trading-bot[dev,ml,web]"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
trading-bot = "trading_bot.main:main"
ai-trading-bot = "trading_bot.scripts.run_ai_bot:main"
mcp-server = "trading_bot.scripts.start_mcp_server:main"
analyze-performance = "trading_bot.scripts.analyze_performance:main"
monitor-bot = "trading_bot.scripts.monitor_bot:main"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "binance.*",
    "ta.*", 
    "pandas_ta.*",
    "telegram.*",
    "feedparser.*",
    "textblob.*",
    "newspaper.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
